pyproject.toml
[project]
name = "genzoo"
version = "0.1.0"
description = "Pipeline for scalable generation of realistic 3D animal pose and shape training data, enabling state-of-the-art estimation models."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "diffusers>=0.35.1",
    "einops>=0.8.1",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "mmcv-lite>=2.0.0rc4,<2.2.0",
    "mmdet>=3.3.0",
    "numpy>=2.3.3",
    "omegaconf>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "protobuf>=6.32.1",
    "pyrender>=0.1.45",
    "pyrootutils>=1.0.4",
    "pytorch-lightning>=2.5.5",
    "scikit-image>=0.25.2",
    "sentencepiece>=0.2.1",
    "smplx>=0.1.28",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
    "trimesh>=4.8.2",
    "wandb>=0.22.0",
    "webdataset>=1.0.2",
    "yacs>=0.1.8",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupyterlab>=4.4.7",
    "notebook>=7.4.5",
]
